save filed

+++ TestGraph/Components/IElecComponent.cs
@@ -4,15 +4,15 @@ namespace TestGraph.Components;
 
 public class ElecComponent : Vertice
 {
-    public int Resistance { get; set; }
+    public double Resistance { get; set; }
     
-    public ElecComponent(string name, int resistance) : base(name)
+    public ElecComponent(string name, double resistance) : base(name)
     {
         if (resistance < 0)
             throw new ArgumentException("Resistance of a component cannot be negative");
         Resistance = resistance;
     }
-    public ElecComponent(string name, List<Vertice> adjacentComponents, int resistance) : base(name, adjacentComponents)
+    public ElecComponent(string name, List<Vertice> adjacentComponents, double resistance) : base(name, adjacentComponents)
     {
         if (resistance < 0)
             throw new ArgumentException("Resistance of a component cannot be negative");
